Renamed to pain (click for explanation).
I wanted to keep this in so everyone can look at the struggles of getting 2 software serial devices to work together while also having problems with the APC220 module not going beyond 1 Hz, basically, a giant rabbit hole of interrupt transmitting and once in a while losing the send data because of interrupting the GPS loop.  (While also minimizing dynamic memory, because the SD card was also unstable, in hindsight I could have done a bit more optimization though)

At least this code worked so we will keep it as a trophy I guess. I used some of it as a reference for writing the new code, I almost wanted to delete but I was like no, I learned too much while working on this.